+++ lib/rrdSupport.py
@@ -106,7 +106,7 @@ class BaseRRDSupport:
 
         # make the start time to be aligned on the rrd_step boundary
         # This is needed for optimal resoultion selection 
-        start_time = (long(time.time() - 1)/rrd_step) * rrd_step 
+        start_time = (int(time.time() - 1)/rrd_step) * rrd_step 
         #print (rrdfname,start_time,rrd_step)+rrd_ds
         args = [str(rrdfname), '-b', '%li' % start_time, '-s', '%i' % rrd_step]
         for rrd_ds in rrd_ds_arr:
@@ -263,7 +263,7 @@ class BaseRRDSupport:
         For more details see
           http://oss.oetiker.ch/rrdtool/doc/rrdcreate.en.html
         """
-        now = long(time.time())
+        now = int(time.time())
         start = ((now-period)/rrd_step)*rrd_step
         end = ((now-1)/rrd_step)*rrd_step
         return self.rrd2graph(fname, rrd_step, ds_name, ds_type, start, end, width, height, title, rrd_files, cdef_arr, trend, img_format)
@@ -396,7 +396,7 @@ class BaseRRDSupport:
         For more details see
           http://oss.oetiker.ch/rrdtool/doc/rrdcreate.en.html
         """
-        now = long(time.time())
+        now = int(time.time())
         start = ((now-period)/rrd_step)*rrd_step
         end = ((now-1)/rrd_step)*rrd_step
         return self.rrd2graph_multi(fname, rrd_step, start, end, width, height, title, rrd_files, cdef_arr, trend, img_format)
